+++ lib.commonjs/providers/provider-jsonrpc.js
@@ -64,7 +64,8 @@ const defaultOptions = {
     batchStallTime: 10,
     batchMaxSize: (1 << 20),
     batchMaxCount: 100,
-    cacheTimeout: 250
+    cacheTimeout: 250,
+    pollingInterval: 4000
 };
 // @TODO: Unchecked Signers
 class JsonRpcSigner extends abstract_signer_js_1.AbstractSigner {
@@ -283,11 +284,7 @@ class JsonRpcApiProvider extends abstract_provider_js_1.AbstractProvider {
         }, stallTime);
     }
     constructor(network, options) {
-        const superOptions = {};
-        if (options && options.cacheTimeout != null) {
-            superOptions.cacheTimeout = options.cacheTimeout;
-        }
-        super(network, superOptions);
+        super(network, options);
         this.#nextId = 1;
         this.#options = Object.assign({}, defaultOptions, options || {});
         this.#payloads = [];
@@ -740,6 +737,12 @@ class JsonRpcApiProvider extends abstract_provider_js_1.AbstractProvider {
     }
 }
 exports.JsonRpcApiProvider = JsonRpcApiProvider;
+// @TODO: remove this in v7, it is not exported because this functionality
+// is exposed in the JsonRpcApiProvider by setting polling to true. It should
+// be safe to remove regardless, because it isn't reachable, but just in case.
+/**
+ *  @_ignore:
+ */
 class JsonRpcApiPollingProvider extends JsonRpcApiProvider {
     #pollingInterval;
     constructor(network, options) {