+++ lib.esm/abi/fragments.js
@@ -18,6 +18,8 @@ function setify(items) {
     items.forEach((k) => result.add(k));
     return Object.freeze(result);
 }
+const _kwVisibDeploy = "external public payable";
+const KwVisibDeploy = setify(_kwVisibDeploy.split(" "));
 // Visibility Keywords
 const _kwVisib = "constant external internal payable private public pure view";
 const KwVisib = setify(_kwVisib.split(" "));
@@ -1021,7 +1023,7 @@ export class ConstructorFragment extends Fragment {
         else if (obj instanceof TokenString) {
             consumeKeywords(obj, setify(["constructor"]));
             const inputs = consumeParams(obj);
-            const payable = !!consumeKeywords(obj, setify(["payable"])).has("payable");
+            const payable = !!consumeKeywords(obj, KwVisibDeploy).has("payable");
             const gas = consumeGas(obj);
             consumeEoi(obj);
             return new ConstructorFragment(_guard, "constructor", inputs, payable, gas);