+++ lib.esm/providers/provider-fallback.js
@@ -230,8 +230,8 @@ export class FallbackProvider extends AbstractProvider {
      *  If a [[Provider]] is included in %%providers%%, defaults are used
      *  for the configuration.
      */
-    constructor(providers, network) {
-        super(network);
+    constructor(providers, network, options) {
+        super(network, options);
         this.#configs = providers.map((p) => {
             if (p instanceof AbstractProvider) {
                 return Object.assign({ provider: p }, defaultConfig, defaultState);
@@ -242,7 +242,15 @@ export class FallbackProvider extends AbstractProvider {
         });
         this.#height = -2;
         this.#initialSyncPromise = null;
-        this.quorum = 2; //Math.ceil(providers.length /  2);
+        if (options && options.quorum != null) {
+            this.quorum = options.quorum;
+        }
+        else {
+            this.quorum = Math.ceil(this.#configs.reduce((accum, config) => {
+                accum += config.weight;
+                return accum;
+            }, 0) / 2);
+        }
         this.eventQuorum = 1;
         this.eventWorkers = 1;
         assertArgument(this.quorum <= this.#configs.reduce((a, c) => (a + c.weight), 0), "quorum exceed provider wieght", "quorum", this.quorum);
@@ -314,7 +322,7 @@ export class FallbackProvider extends AbstractProvider {
         // Shuffle the states, sorted by priority
         const allConfigs = this.#configs.slice();
         shuffle(allConfigs);
-        allConfigs.sort((a, b) => (b.priority - a.priority));
+        allConfigs.sort((a, b) => (a.priority - b.priority));
         for (const config of allConfigs) {
             if (config._lastFatalError) {
                 continue;